+++ AddOns/BetterBlizzFrames/temp_era/BetterBlizzFrames.lua
@@ -521,7 +521,7 @@ ClickthroughFrames:RegisterEvent("PLAYER_REGEN_DISABLED")
 
 function BBF.PlayerElite(mode)
     local playerElite = PlayerFrameTexture
-    local bigHealthbars = BetterBlizzFramesDB["biggerHealthbars"]
+    local bigHealthbars = BetterBlizzFramesDB["biggerHealthbars"] and not BetterBlizzFramesDB.biggerHealthbarsNoPlayer
     local hideMana = BetterBlizzFramesDB.hidePlayerManabar
 
     -- Set Elite style according to value
@@ -874,8 +874,8 @@ local function IsFontFileValid(path)
 		return false
 	end
 
-	local ok, result = pcall(FontValidatorString.SetFont, FontValidatorString, path, 12, "")
-	if not ok or result == false then
+	local ok = pcall(FontValidatorString.SetFont, FontValidatorString, path, 12, "")
+	if not ok then
 		return false
 	end